Day 13, the algorithm may contain traces of AI...

+++ src/main/kotlin/Coordinates.kt
@@ -1,10 +1,11 @@
-data class Coordinates(val x: Int, val y: Int) {
+data class Coordinates(val x: Long, val y: Long) {
     fun moveTo(direction: Direction) = this.copy(x = x + direction.x, y = y + direction.y)
 
     operator fun plus(other: Coordinates) = Coordinates(x + other.x, y + other.y)
     operator fun minus(other: Coordinates) = Coordinates(x - other.x, y - other.y)
 
     fun neighbours() = Direction.entries.map(this::moveTo)
+
 }
 
 enum class Direction(val x: Int, val y: Int) {
